TITANIC SURVIVAL PREDICTION
This project aims to predict the survival of passengers aboard the Titanic using machine learning techniques. The sinking of the RMS Titanic in 1912 is one of the most infamous shipwrecks in history. The tragic event resulted in the loss of many lives, but various factors influenced the survival chances of individuals.

In this project, we leverage a dataset containing information about Titanic passengers such as their age, gender, ticket class, and whether they survived or not. We explore the data, perform data preprocessing, feature engineering, and then train machine learning models to predict passenger survival.

Through this predictive modeling task, we hope to gain insights into the factors that contributed to survival during the Titanic disaster. Additionally, this project serves as a learning opportunity to understand and apply fundamental concepts of data analysis, machine learning, and model evaluation.